// frontend/src/generated/model/place.ts
/**
 * backend
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PlaceDataInterface } from './placeDataInterface';


export interface Place { 
    /**
     * A class representation of the BSON ObjectId type.
     */
    _id?: string;
    name?: string;
    type: number;
    longitude: number;
    latitude: number;
    email?: string | null;
    website?: string | null;
    phone?: string | null;
    schedule?: any | null;
    isPaid?: boolean | null;
    price?: string | null;
    address?: string | null;
    data?: PlaceDataInterface | null;
    categories?: Array<string>;
}